refactor: move sceneframe to context folder
