+++ packages/client/builder/src/xform/axes/buildAxis.ts
@@ -10,7 +10,7 @@ import {
 	ViewSize,
 } from '@chart-parts/interfaces'
 import { buildMark } from '@chart-parts/scenegraph'
-import { SceneFrame } from '../SceneFrame'
+import { SceneFrame } from '../context/SceneFrame'
 import { components } from './components'
 import { SGMarkAny } from '../processNode'
 import { AxisSpace } from '../interfaces'
